refactor udfgenerator: move FunctionParts and friends to ast.py
